[keyserver] Stop sending incorrect macOS rescinds
Summary:
[ENG-4523](https://linear.app/comm/issue/ENG-4523/stop-sending-incorrect-macos-rescinds)

Because of handling of older deliveries with only `iosID`, and because we were always setting parsedDeliveries' platform to `'ios'` we were incorrectly trying to send rescinds for macOS notifications. This diff makes it so we are setting the platform correctly, and the parsed deliveries for macos are filtered later in the code.

Test Plan: - Check that after displaying a notification no rescinds for macOS are sent

Reviewers: kamil, tomek, marcin, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8699