[desktop] Upgrade node-abi
Summary:
[ENG-4537](https://linear.app/comm/issue/ENG-4537/windows-ci-failing-after-electron-bump)

The version of the `node-abi` that we had in lockfile didn't have entries for the new electron versions. Upgrading it to this new version should help with the windows CI build failures (but I couldn't reproduce them myself so we will see). I've confirmed that the newest version has entries for electron v25.

I tried 3 ways to update the version:
- `yarn upgrade node-abi` did nothing, probably because `node-abi` isn't a direct dependency
- Following [instructions from the main electron maintainer](https://github.com/electron/rebuild/issues/1020#issuecomment-1162234866) run `npx uuaw node-abi` which updated the lockfile to the version in this diff
- Tried manually removing entries for `node-abi` from the lockfile and running `yarn install` which also ended in the lockfile version that's in this diff (matched the previous method)

Test Plan: `yarn cleaninstall`

Reviewers: atul, ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8682