[native] correctly measure multimedia messages that have an inline engagement
Summary:
This diff handles correctly measuring multimedia messages. Multimedia needed to be handled a little bit differently since we use the image height and the composed message width to determine the `contentHeight` and we actually don't perform a measurement on multimedia messages.

This means that for multimedia messages we only need to measure the inline engagement and then in `multimediaMessageItemHeight` we can add the `contentHeight` w/ the height of the inline engagement we get from the `NodeHeightMeasurer`.

Depends on D8632

Test Plan:
I tested modifying the height of the inline engagement for multimedia messages, and when I did this I no longer got incorrect height measurement logs from the console. I performed the same tests outlined in the test plan D8632 (except edit label test, but you can't edit multimedia message)

Before:
{F654799}

After:
{F654800}

Reviewers: atul, kamil, ashoat, tomek

Reviewed By: tomek

Subscribers: rohan, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8634