[web] introduce functionality to send a reaction by clicking on the reaction pill
Summary:
To improve the user experience of the `InlineEngagement` component we are updating the clicking behavior for our reaction pills. Before whenever a user would click on a reaction, the `MessageReactionsModal` would pop up, but now if a user clicks a reaction pill they will react/unreact to that message depending on whether or not they have already reacted to that message already. To view the MessageReactionsModal the user now needs to hover on the reaction pill and click on the popup (a similar pattern with other chat apps such as discord and slack), and that functionality will be handled in subsequent diffs

Depends on D8807

Test Plan:
Please watch the demo below:

{F688372}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8808