[web] show which reactions the viewer has already selected in the inline engagement
Summary: To improve the user experience of the inline engagement, we should show/indicate to the user which reactions they have already selected. We indicate this to the user by adding a border to the reaction pill.

Test Plan: Confirmed the width/height of the reaction pills are all the same regardless if it has been selected by the viewer or not

Reviewers:

Subscribers: