[keyserver] Convert ids in rescinds
Summary:
[ENG-4569](https://linear.app/comm/issue/ENG-4569)

Rescinds also need to be converted. To do that we need to pass `stateVersion` in the delivery, alongside the `codeVersion` and using it convert the rescinds.

Test Plan: Checked that notification were rescinded on macOS and Android, after displaying them on web app. Checked if the `rescinded` column in the `notification` table was set to 1, and a new rescind row was created. Removed `stateVersion` from one of the rescinds manually and checked if the notification was still rescinded.

Reviewers: marcin!, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8706