[keyserver] Convert ids in rescinds
Summary:
[ENG-4569](https://linear.app/comm/issue/ENG-4569)

Rescinds also need to be converted. To do that we need to pass `stateVersion` in the delivery, alongside the `codeVersion` and using it convert the rescinds.

Test Plan:
Checked that notification were rescinded on macOS and Android, after displaying them on web app. Checked if the `rescinded` column in the `notification` table was set to 1, and a new rescind row was created.

Done this for both `stateVersion < 43` (the ids aren't converted) and `stateVersion >= 43` (the ids were converted).
Tested variant when the notification was sent to `stateVersion < 43` but rescind was sent to `stateVersion >= 43` (in this case I also checked if the group summary notification was removed).
Also tested if the notifications were rescinded if they sent from before these changes (current master) and so wouldn't have the `stateVersion` in delivery.

Reviewers: marcin, tomek

Reviewed By: marcin, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8706