[GH Actions] Add `gradle.jvmargs` args to `./gradlew bundleRelease`
Summary:
This is kind of a longshot, but noticed that the disk usage issue was accompanied by a stackoverflow issue during some sort of final signing step. This issue doesn't replicate on my fork because I didn't include the keys necessary for signing.

I also noticed that the disk usage issue specified that the error occured when trying to write to some sort of build log file.

Maybe there's a stack overflow issue that's causing a lot of logging that's eating up disk space? It's a stretch, but figure it doesn't hurt to bump `gradle.jvmargs` to possibly reduce change of stack overflow?

Test Plan:
Land and see what happens.

Reviewers:
ashoat, ginsu, marcin, tomek, jon