[keyserver] Include stateVersion when converting notifications
Summary:
[ENG-4579](https://linear.app/comm/issue/ENG-4579)

ID conversion is based on the `stateVersion`. Unfortunately the notif code only handled the `codeVersion` and passed a new `platformDetails` object that didn't contain the `stateVersion` so this diff improves that.

This wasn't noticed earlier because previously conversion was based on `codeVersion` so it was working properly. Additionaly created [ENG-4850](https://linear.app/comm/issue/ENG-4580) to refactor this code to handle the `platformDetails` together in the future.

Test Plan:
- Tested sending notification on iOS, macOS, Android, Web and Windows.
  - On clients with `stateVersion < 43` the ids weren't converted
  - On clients with `stateVersion >= 43` the ids were converted

On mobile clients, the IDs were tested in the JS (not native code.)

Reviewers: tomek, marcin!, atul

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8720