[keyserver] Include stateVersion when converting notifications
Summary:
[ENG-4579](https://linear.app/comm/issue/ENG-4579)

ID conversion is based on the `stateVersion`. Unfortunately the notif code only handled the `codeVersion` and passed a new `platformDetails` object that didn't contain the `stateVersion` so this diff improves that.

This wasn't noticed earlier because previously conversion was based on `codeVersion` so it was working properly. Additionaly created [ENG-4850](https://linear.app/comm/issue/ENG-4580) to refactor this code to handle the `platformDetails` together in the future.

Test Plan:
Tested on Android, iOS, macOS and web that the received notification contained converted ids. I will be able to test windows in the evening and will amend this test plan.

Reviewers: tomek, marcin, atul

Subscribers: