[keyserver][lib][web][native] Move updatesCurrentAsOf field to keyserverSotre
Summary:
issue: https://linear.app/comm/issue/ENG-4458/refactor-updatescurrentasof-field
For now, we are using the id of our keyserver. In the future, actions that change fields in the keyserverStore will include in their payload the id of the keyserver the action is related to. That
id will be used in place of `ashoatKeyserverID`. Thiss will be done during https://linear.app/comm/issue/ENG-4440/refactor-redux-fields-second-iteration-refactor-actions

Test Plan: Ran `yarn flow-all`. Checked with redux devtools that the field gets correctly migrated to keyserverStore

Reviewers: michal, kamil

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8719