[web] Introduce `useUploadAvatarMedia` to `AvatarHooks`
Summary:
This function is similar to `useUploadProcessedMedia()` in `native/avatar-hooks`. It basically lets us upload the `File` we receive from image `input`.

We consume `useUploadAvatarMedia` in `EditUserAvatarMenu` to upload the selected image. In subsequent diffs, we will update `onImageSelected` to handle the setting of the image avatar as well.

Test Plan:
1. Select "Select image" from `EditUserAvatarMenu`
2. Select an image from the file explorer
3. Check the `uploads` table and observe that the image was uploaded:

Reviewers:
ashoat, ginsu