[web] Fix small thread list scrolling issue
Summary:
Just noticed that if you put your mouse cursor on the thread list in a place where the scrollbar normally displays you won't be able to scroll because there will another top-level scrollbar (for some reason on production I can only reproduce it in cases where the threads don't fill the whole list).

This was introduced when I landed [ENG-3976](https://linear.app/comm/issue/ENG-3976) because the thread list container now contains the `react-window` list and the scrollbars overlap each other. This is a quick fix that sets `overflow: hidden` on the container (we don't want to scroll it, the scrolling is handled by the inner list).

Test Plan:
Try the reproduction mentioned in the summary and check if the scrolling works and the second scrollbar doesn't appear

Reviewers: inka, patryk, rohan, ginsu

Subscribers: