[keyserver][lib][native][web] Refactor currentAsOf field in messageStore
Summary:
issue: https://linear.app/comm/issue/ENG-4461/refactor-currentasof-field-in-message-store
`currentAsOf` is a timestamp. The client claims to have all messages that are interesting to it up until this point in time. It the socket needs to be re-initialiezed, it uses this value in the initial message to the server, to fetch
messages that came after. It is updated when the user logs in or recievies more messages. In the second case it is updated to the timestamp of the newest message the client has ever seen. So it makes sense to leave `currentAsOf` in
`messageStore`, because it is inferred from the value of `messageStore` when it updates.

Test Plan: Ran `yarn flow-all`. Tested that the migration on native works correctly. Checked that no errors are thrown from validators when web app is opened.

Reviewers: michal, kamil

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8728