[native] Fix encrypted media upload failures
Summary:
Fixes [[ https://linear.app/comm/issue/ENG-4566/encrypted-media-fails-to-upload-on-ios | ENG-4566 ]]. Issue details are in the task.

Previously, temporary encrypted file was created in the same directory as original. This diffs moves it to the app cache dir.

Test Plan:
Ensured that encrypted media are successfully uploaded on both iOS (physical device) and Android. Also ensured that it is properly disposed after upload.

Console logged modified paths to confirm that they're correctly costructed in the app cache dir:

```
lang=json
{
  "destinationPath": "/Users/barthap/Library/Developer/CoreSimulator/Devices/4BFD741A-64AB-4720-8125-3874478D8B4B/data/Containers/Data/Application/64F264B3-A8CF-4237-9B72-A2EFCE8B8A73/tmp/IMG_0011.dat",
  "destinationURI":"file:///Users/barthap/Library/Developer/CoreSimulator/Devices/4BFD741A-64AB-4720-8125-3874478D8B4B/data/Containers/Data/Application/64F264B3-A8CF-4237-9B72-A2EFCE8B8A73/tmp/IMG_0011.dat",
  "originalFilename":"IMG_0011.PNG",
  "targetFilename":"IMG_0011.dat",
  "uri":"file:///Users/barthap/Library/Developer/CoreSimulator/Devices/4BFD741A-64AB-4720-8125-3874478D8B4B/data/Media/DCIM/100APPLE/IMG_0011.PNG"
}
```

Reviewers: atul, ashoat, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8738