[native] Show loading spinner during getOlmSessionInitializationData server call
Summary: This diff addresses [ENG-4604](https://linear.app/comm/issue/ENG-4604/delay-before-spinner-starts-spinning-on-mobile-log-in).

Test Plan:
{F679488}

Note that there is a bit of a stutter when the first `loadingSelector` finishes, since the second one doesn't overlap with the first. I decided that this was okay for now... we'll end up needing to change things here pretty soon anyways, since these keys will get fetched from the identity service instead of the keyserver in the near future.

Reviewers: atul, marcin, ginsu

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8762