[web] Hide reply tooltip button for multimedia
Summary: Just took a second to change this, we don't want to check whether the message is composible since that includes both text and multimedia. Now, if the message is not text we will not show the tooltip
icon.

https://linear.app/comm/issue/ENG-4520/reply-button-appears-for-images-on-web

Test Plan: See below and after to verify that the reply button no longer shows for multimedia

Reviewers: atul, ginsu

Subscribers: