[native] Introduce `useNativeSetThreadAvatar` and use throughout `native`
Summary:
Thread equivalent of D8341

`setThreadAvatar` previously had a call to `displayFailureAlert`, however, this function is only relevant on `native` since we don't display alerts on `web`.

As part of making the provider platform-agnostic, we "lift up" error handling/alert displaying to `avatar-hooks`.

We replace usages of `setThreadAvatar` on `native` with `nativeSetUserAvatar`.

Test Plan: The modified flows on `native` continue to work as expected.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8757