[keyserver] Improve Flow types in update-creator
Summary:
We had a type error and Flow was unable to detect that. This diff makes it so the type error would be reported and replaces `false` with `null` to satisfy type system. In `fetchMessageInfos` and `parseMessageSelectionCriteria` we're using `threadCursors[threadID]` in `if` conditions so it is safe to do that replacement.

https://linear.app/comm/issue/ENG-4592/improve-flow-types-in-update-creator

Test Plan: Console log `messageInfoResult` before and after change and check if the result remains correct.

Reviewers: kamil, bartek, inka