[native] Create a LinkedDevicesHeaderRightButton that will open the QR scanner
Summary:
We want to allow the react navigation header right button to be pressed in order to open the QR code scanner for the new login flow. The `onPress` action will come in a following diff, this just adds the header right component to the new LinkedDevices screen.

I know I could have inlined the `headerRight` component probably in `profile.react.js`, but since I plan to add some additional checks (i.e. if the device is on a simulator) when the header right is pressed before opening the QR code scanner, I thought it would be cleaner extracting it to its own component file.

Part of [[ https://linear.app/comm/issue/ENG-4476/add-a-linked-devices-section-on-the-app-for-primary-devices | ENG-4476 ]]

Depends on D8764

Test Plan:
Confirmed that the header right appears in the LinkedDevices screen

{F682981}

Reviewers: atul, ginsu, varun, tomek

Reviewed By: ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8769