[native] fix forcing the incorrect height for multimedia messages
Summary:
This diff is part 3/3 to fix forcing the incorrect height for messages with inline engagement in prod.

This [[ https://linear.app/comm/issue/ENG-4606/fix-forcing-incorrect-height-for-messages-with-inline-engagement-in | linear ]] comment explains this issue and what I did to address this

> Since item.contentHeight is now the height of the inner text message AND the height of the inline engagement we are now forcing the wrong height for the inner text message component. An idea I want to explore to address this is to force the height higher up in the component tree (somewhere like ComposedMessage) where both the "inner message" and the inline engagement is rendered so that when we force the height withitem.contentHeight, it is correct

Depends on D8798

Test Plan:
Please see demo video where I force and unforce the correct height of the multimedia messages and it did not change visually (outside of the green background) and I did not get any error logs talking about incorrect height measurements, I also purposely force an incorrect height (which resulted in some logs about incorrect height measurements.

Also please note for the video, I set a green background so that it is easier to visually see what is happening, but these are only for the demo videos

With the correct height measurement:
{F686605}

With the incorrect height measurement:
{F686622}

Here is also a demo video to show that there was no regression by removing the `expand` style (I pressed and was able to view the the full screen image for multimedia messages with both a single image and multiple images):
{F689819}

Also here is another demo video to view the the full screen image for multimedia messages with both a single image and multiple images W/O an inline engagement
{F689917}

Reviewers: atul, kamil, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8799