Move `native`-specific `uploadSelectedMedia` from `updateImageThreadAvatar` to `nativeUpdateImageThreadAvatar`
Summary:
We remove the `native`-specific `uploadSelectedMedia` call from `updateImageThreadAvatar` and move it to the `native`-specific `nativeUpdateThreadImageAvatar`.

As of this diff, `updateImageThreadAvatar` now takes `(imageAvatarUpdateRequest: ImageAvatarDBContent)` rather than the `native`-specific `(selection: NativeMediaSelection)`.

This diff also eliminates all usages of `useUploadSelectedMedia`, which means we can remvoe the prop in a subsequent diff at which point `EditThreadAvatarProvider` will be platform-agnostic.

Test Plan: Continue to be able to set image thread avatars on `native` as expected.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8772