[services-lib] Introduce Blob Client
Summary:
Part of [[ https://linear.app/comm/issue/ENG-4597/blob-client-for-services | ENG-4597 ]].

This diff introduces a foundation for blob client in comm-services-lib. The client is hidden behind a crate feature so services that do not need it do not have to depend on it (e.g. feature-flags).

Test Plan: Cargo build in comm-services-lib, blob and feature-flags.

Reviewers: michal, jon

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8777