[native] Set up the BarCodeScanner to scan QR codes
Summary:
This diff introduces the usage of the Expo `BarCodeScanner` when the header right 'Add' is clicked on the linked devices screen.

The [[ https://docs.expo.dev/versions/latest/sdk/bar-code-scanner/ | Expo docs ]] were closely followed in setting this up.

For now, the `onConnect` will not do anything besides alert the user that the qr code has been scanned with the data encoded, but down the line this will be changed to read the secondary devices' keys encoded in the qr code.

Addresses both [[ https://linear.app/comm/issue/ENG-4475/use-the-expo-barcodescanner-library-to-read-qr-codes-from-within-the | ENG-4475 ]] and [[ https://linear.app/comm/issue/ENG-4477/alert-the-primary-device-when-the-qr-code-is-scanned-with | ENG-4477 ]].

Depends on D8771

Test Plan:
Confirmed that the scanner opened if I granted camera permissions, and also that it blocked launching the scanner if I denied permissions.

Accepting permissions:
{F688225}

Denying permissions:
{F688235}

Reviewers: atul, ginsu, tomek, varun

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8805