[native] Set up the BarCodeScanner to scan QR codes
Summary:
This diff introduces the usage of the Expo `BarCodeScanner` when
the header right 'Add' is clicked on the linked devices screen.

The [[ https://docs.expo.dev/versions/latest/sdk/bar-code-scanner/ | Expo
docs ]] were closely followed in setting this up.

For now, the `onConnect` will not do anything besides alert the user that
the qr code has been scanned with the data encoded, but down the line this
will be changed to read the secondary devices' keys encoded in the qr
code.

Test Plan:
Confirmed that the scanner opened if I granted camera
permissions, and also that it blocked launching the scanner if I denied
permissions.

Reviewers: atul, ginsu, tomek, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8805