[native] Prevent simulators from attempting to open the barcode scanner
Summary:
Launching the barcode scanner on a simulator crashes the entire app (this is because it is not a [[ https://docs.expo.dev/versions/latest/sdk/bar-code-scanner/#platform-compatibility | compatible platform ]].

We should prevent this from being accessed in the simulator.

Addresses [[ https://linear.app/comm/issue/ENG-4647/prevent-simulator-from-opening-the-scanner | ENG-4647 ]]

Depends on D8805

Test Plan:
Confirmed that the scanner no longer is launched when on a simulator.

{F688295}

Reviewers: atul, ginsu, tomek, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8806