[keyserver][lib][native] Add urlPrefix to keyserverStore
Summary:
issue: https://linear.app/comm/issue/ENG-4459/refactor-urlprefix-field
`urlPrefix` needs to be moved to `keyserverStore`. In next diffs I will start using this value and remove the `urlPrefix` from where it was previously (`BaseAppState`).
For now we are updating the urlPrefix of "Ashoat's keyserver" when we want to connect to a different keyserver. During https://linear.app/comm/issue/ENG-4130/settings-panel-for-keyserver-selection some logic will be written, that will allow to add a new keyserver instead. Then, `setUrlPrefix` will be used only if the keyserver changes it's url, not as a way of connecting to a //different// keyserver

Test Plan: Ran `yarn flow-all`

Reviewers: michal, kamil

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8813