[web] introduce calculateReactionTooltipSize function
Summary:
This diff introduces a function to calculate the size of the reaction tooltip. In order to build a tooltip, we need to first know the size, so this function will handles that for our new reaction tooltip.

Here is a screenshot of the reaction for context:
{F694576}

Depends on D8858

Test Plan:
logged out the values of `width` and `height` and confirmed that they matched the size of the tooltip that was being rendered

{F694587}

{F694583}

{F694586}

{F694585}

{F694582}

{F694584}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8857