[native] fix RetrySend not visible when message fails
Summary:
This diff fixes the regression where the `RetrySend` component doesn't appear when a message fails. This was due to me not considering this element when we were forcing the height for composed messages.

To fix this I am now forcing the height in `Message` and forwarding down `viewStyle` to `TextMessage`, `MultimediaMessage`, and `RobotextMessage`. This will end up setting the height in exactly the same places that it's being set in the previous revision, but now with the added benefit of localizing the height-warning logic in the same place that the height-forcing logic. We also actually already have a function (`messageItemHeight`) that calculates the height of this root view for all three of the message types so we can just reuse that here.

Test Plan:
Please see demo video where I force and unforce the height of the composed message and the `RetrySend` component remains visible. I also confirmed that I did not get any error logs talking about incorrect height measurement with this change for any of the three message types.

Before:
{F692535}
{F692546}

After (the green background is just to show that the simulator is being updated when forcing/unforcing the height):
{F693639}

Reviewers: atul, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8839