[backup] Introduce backup error
Summary:
[ENG-4380](https://linear.app/comm/issue/ENG-4380/initial-version-of-backup-service)

Catch-all error for all errors that will be able to come up during handling of backup requests. For now it only has errors for blob client, but it will be expanded in the future diffs.

Test Plan:
Create a service:
```
 .service(web::resource("/hello").route(web::get().to(
        |blob: web::Data<BlobServiceClient>| async move {
          let _ = blob.get("asdasd").await.map_err(BackupError::from)?;
          Ok::<&str, BackupError>("world")
        },
      )))
```
Test a few different scenarios: blob error not available (not running), blob not existing

Reviewers: bartek, patryk, bartek

Subscribers: