[keyserver] Use Set as an argument of fetchEntryInfosByID
Summary:
Refactor `fetchEntryInfosByID` so that it accepts a set of ids. Something similar could be done for `fetchKnownUserInfos` but it is used in a lot of places so it would be expensive.
Also, replace ids arrays with sets in `checkState` - that makes sense on its own, but also is beneficial because it makes interfaces of specs more unified.

https://linear.app/comm/issue/ENG-4654/unify-interfaces-in-session-utils

Test Plan:
Change some entries and users in the keyserver db, increase state check frequency, start the app and check if the state got fixed.
Add console logs in `fetchEntryInfosByID` and check if the correct result is returned.
Check if correct data is fetched in `checkState`.

Reviewers: kamil, bartek, inka

Reviewed By: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8925