[keyserver] Return object from fetchEntryInfosByID
Summary:
All the others state syncs use objects instead of array when fetching entries, so it makes sense to make this consistent. Additionally, it is more efficient to access by id instead of iterating over array.

https://linear.app/comm/issue/ENG-4654/unify-interfaces-in-session-utils

Depends on D8925

Test Plan: Modify a couple of calendar entries in the keyserver db and check if the state is correctly fixed.

Reviewers: kamil, bartek, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8926