[web] Introduce `ThreadEmojiAvatarSelectionModal`
Summary:
Enables setting of emoji avatars on `web`. As of this diff all varieties of thread avatar can be set from `web`. We just need to make a few layout tweaks in `ThreadSettingsModalGeneralTab` in order to get this ready to ship.

Test Plan:
Works as expected: [will insert video]

Reviewers:
ashoat, ginsu