[native/web] change title of MessageReactionsModal to All reactions
Summary:
I am wrapping up the inline engagement redesign work right now, but didn't have the chance to work on redesigning the `MessageReactionsModal`. With the new inline engagement going to our old `MessageReactionsModal` felt a little jarring since the modal provides all the reactions to you and not just the reaction specific ones.

{F694719}

To make it a little less jarring I changed the title from "Reactions" to "All reactions" and I am hoping that this will make it more clear to the user that we are not showing the list of users for that specific reactions.

https://linear.app/comm/issue/ENG-3787/implement-a-redesign-for-message-reactions-modal is the task for this, and I am hoping to get to it whenever I have a free cycle

Test Plan:
Please see screenshots

native:
{F694715}

web:
{F694716}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8861