[client-backup] implement C++ function for computing `BackupKey`
Summary:
[ENG-4512](https://linear.app/comm/issue/ENG-4512#comment-75b01dad).

Note:
This is call to Rust, but it's related to cryptography, so I believe it should be executed on `cryptoThread` which is spawned via `CommCoreModule` - this is why I placed this here, not in `CommRustModule`. In the future, we might want to introduce new module to crypto-related things.

Depends on D8885

Test Plan: Call `commCoreModule.computeBackupKey`.

Reviewers: michal, marcin, varun

Reviewed By: marcin, varun

Subscribers: varun, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8886