[client-backup] implement C++ functions for encoding/decoding string in `utf-8`
Summary:
[ENG-4503](https://linear.app/comm/issue/ENG-4503/encryptdecrypt-userdata-with-backupdatakey-ct2).

To encrypt/decrypt `UserData` we need to have it in bytes (Uint8Array in JS or uint8_t[] in C++). From JS level it's not that easy because standard TextEncoder from JS is not available in react-native, so we're using JSI capabilities to convert this.

Depends on D8887

Test Plan:
```
      const str = '!ABC+** /==🤬';
      console.log(str);
      const bytes = commUtilsModule.encodeString(str);
      const arr = new Uint8Array(bytes);
      console.log(arr);
      const str2 = commUtilsModule.decodeString(arr.buffer);
      console.log(str2);
      console.log(str === str2);
```

Reviewers: michal, marcin