[native] Add a temporary sign in via QR code button on native
Summary:
While we're still using the traditional form to log in on `native`, I'm hiding the QR code login flow behind a new login button (similar to what's done with the register (new) button). This is only
shown in development. The next diff will implement the screen that appears when clicking it.

I followed an approach similar to D7495

Addresses [[ https://linear.app/comm/issue/ENG-4649/add-a-temporary-sign-in-via-qr-code-button-on-native | ENG-4649 ]]

Depends on D8836

Test Plan: Confirmed that the button appears only in development

Reviewers: atul, ginsu, tomek, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8870